ia_backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
langchain-google-genai
langchain
PyMuPDF
faiss-cpu
numpy
python-multipart